+++ lib/ts/recipe/emailpassword/utils.ts
@@ -36,6 +36,7 @@ import {
     defaultValidate,
 } from "./validators";
 import { normaliseAuthRecipeModuleConfig } from "../authRecipeModule/utils";
+import { RecipeImplementation, RecipeInterface } from "./";
 
 export function normaliseEmailPasswordConfig(config: Config): NormalisedConfig {
     const signInAndUpFeature: NormalisedSignInAndUpFeatureConfig = normaliseSignInAndUpFeature(
@@ -61,10 +62,26 @@ export function normaliseEmailPasswordConfig(config: Config): NormalisedConfig {
             config.resetPasswordUsingTokenFeature
         );
 
+    let override: {
+        functions: (originalImplementation: RecipeImplementation) => RecipeInterface;
+    } = {
+        functions: (originalImplementation: RecipeImplementation) => originalImplementation,
+    };
+
+    if (config !== undefined && config.override !== undefined) {
+        if (config.override.functions !== undefined) {
+            override = {
+                ...override,
+                functions: config.override.functions,
+            };
+        }
+    }
+
     return {
         ...normaliseAuthRecipeModuleConfig(config),
         signInAndUpFeature,
         resetPasswordUsingTokenFeature,
+        override,
     };
 }