+++ lib/ts/recipe/recipeModule/index.ts
@@ -85,5 +85,5 @@ export default abstract class RecipeModule<T, S, R, N extends NormalisedConfig<T
 
     abstract getFeatures(): RecipeFeatureComponentMap;
 
-    abstract getFeatureComponent(componentName: string): JSX.Element;
+    abstract getFeatureComponent(componentName: string, props: any): JSX.Element;
 }