+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     -   Uses redirectToAuth everywhere instead of calling redirect manually.
     -   Adds additional props / config to Session and EmailVerification recipe to make them more isolated
 -   In session recipe, we remove `setAuth0API` and `getAuth0API`
+-   Removes a global pre API hook per recipe. Pre API hooks now need to be added per function in the recipe interface.
 
 ## [0.13.0] - 2021-05-11